[Fixes #6957] Duplicated registered attributes in GeoNode - and Update sttributes can't fix it (#6958)
