+++ .circleci/config.yml
@@ -123,7 +123,7 @@ workflows:
           name: geonode_test_suite_smoke
           load_docker_cache: true
           save_docker_cache: true
-          test_suite: coverage run --branch --source=geonode manage.py test -v 3 --keepdb geonode.tests.smoke
+          test_suite: coverage run --branch --source=geonode manage.py test -v 3 --keepdb geonode.tests.smoke geonode.tests.test_message_notifications geonode.tests.test_rest_api geonode.tests.test_search geonode.tests.test_utils
       - build:
           name: geonode_test_suite
           load_docker_cache: true
