+++ geonode/tests/utils.py
@@ -17,11 +17,9 @@
 # along with this program. If not, see <http://www.gnu.org/licenses/>.
 #
 #########################################################################
-
 from geonode.tests.base import GeoNodeBaseTestSupport
 
 import os
-import copy
 import time
 import base64
 import pickle
@@ -46,17 +44,12 @@ from requests.packages.urllib3.util.retry import Retry
 from requests_toolbelt.multipart.encoder import MultipartEncoder
 
 from django.core import mail
-from django.conf import settings
-from django.db.models import signals
 from django.urls import reverse
-from django.core.management import call_command
+from django.conf import settings
 from django.contrib.auth import get_user_model
-from django.contrib.gis.geos import Polygon
 from django.test.client import Client as DjangoTestClient
 
 from geonode.maps.models import Layer
-from geonode.geoserver.helpers import set_attributes
-from geonode.geoserver.signals import geoserver_post_save
 from geonode.notifications_helper import has_notifications, notifications
 
 logger = logging.getLogger(__name__)
@@ -324,56 +317,6 @@ def check_layer(uploaded):
     assert isinstance(uploaded, Layer), msg
     msg = ('The layer does not have a valid name: %s' % uploaded.name)
     assert len(uploaded.name) > 0, msg
-
-
-class TestSetAttributes(GeoNodeBaseTestSupport):
-
-    def setUp(self):
-        super(TestSetAttributes, self).setUp()
-        # Load users to log in as
-        call_command('loaddata', 'people_data', verbosity=0)
-
-    def test_set_attributes_creates_attributes(self):
-        """ Test utility function set_attributes() which creates Attribute instances attached
-            to a Layer instance.
-        """
-        # Creating a layer requires being logged in
-        self.client.login(username='norman', password='norman')
-
-        # Disconnect the geoserver-specific post_save signal attached to Layer creation.
-        # The geoserver signal handler assumes things about the store where the Layer is placed.
-        # this is a workaround.
-        disconnected_post_save = signals.post_save.disconnect(geoserver_post_save, sender=Layer)
-
-        # Create dummy layer to attach attributes to
-        _l = Layer.objects.create(
-            name='dummy_layer',
-            bbox_polygon=Polygon.from_bbox((-180, -90, 180, 90)),
-            srid='EPSG:4326')
-
-        # Reconnect the signal if it was disconnected
-        if disconnected_post_save:
-            signals.post_save.connect(geoserver_post_save, sender=Layer)
-
-        attribute_map = [
-            ['id', 'Integer'],
-            ['date', 'IntegerList'],
-            ['enddate', 'Real'],
-            ['date_as_date', 'xsd:dateTime'],
-        ]
-
-        # attribute_map gets modified as a side-effect of the call to set_attributes()
-        expected_results = copy.deepcopy(attribute_map)
-
-        # set attributes for resource
-        set_attributes(_l, attribute_map)
-
-        # 2 items in attribute_map should translate into 2 Attribute instances
-        self.assertEqual(_l.attributes.count(), len(expected_results))
-
-        # The name and type should be set as provided by attribute map
-        for a in _l.attributes:
-            self.assertIn([a.attribute, a.attribute_type], expected_results)
 
 
 if has_notifications:
